Moved DistanceElement classes into own package.
