+++ src/org/vitrivr/cineast/core/data/distance/DistanceElement.java
@@ -1,4 +1,4 @@
-package org.vitrivr.cineast.core.data.score;
+package org.vitrivr.cineast.core.data.distance;
 
 import java.util.Comparator;
 import java.util.List;
@@ -7,8 +7,8 @@ import java.util.stream.Stream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.vitrivr.cineast.core.data.CorrespondenceFunction;
+import org.vitrivr.cineast.core.data.score.ScoreElement;
 import org.vitrivr.cineast.core.util.GroupingUtils;
-import org.vitrivr.cineast.core.util.ReflectionHelper;
 
 public interface DistanceElement {
   String getId();
