+++ src/org/vitrivr/cineast/core/db/ImporterSelector.java
@@ -9,9 +9,9 @@ import org.vitrivr.cineast.core.config.Config;
 import org.vitrivr.cineast.core.config.QueryConfig;
 import org.vitrivr.cineast.core.config.ReadableQueryConfig;
 import org.vitrivr.cineast.core.data.FixedSizePriorityQueue;
+import org.vitrivr.cineast.core.data.distance.DistanceElement;
 import org.vitrivr.cineast.core.data.providers.primitive.FloatTypeProvider;
 import org.vitrivr.cineast.core.data.providers.primitive.PrimitiveTypeProvider;
-import org.vitrivr.cineast.core.data.score.DistanceElement;
 import org.vitrivr.cineast.core.importer.Importer;
 import org.vitrivr.cineast.core.util.distance.FloatArrayDistance;
 import org.vitrivr.cineast.core.util.distance.PrimitiveTypeMapDistanceComparator;
